+++ modules/back-end/src/Api/appsettings.Development.json
@@ -12,7 +12,10 @@
     },
     "WriteTo": [
       {
-        "Name": "Console"
+        "Name": "Console",
+        "Args": {
+          "outputTemplate": "[{Timestamp:HH:mm:ss} {Level:u3}] {SourceContext}: {Message:lj}{NewLine}{Exception}"
+        }
       }
     ]
   },
