+++ modules/evaluation-server/src/Api/Program.cs
@@ -1,15 +1,18 @@
 using Api.Setup;
 using Serilog;
-using Serilog.Events;
-using Serilog.Formatting.Compact;
-using Serilog.Sinks.OpenTelemetry;
 
 try
 {
-    InitializeSerilog();
+    // The initial "bootstrap" logger is able to log errors during start-up. It's completely replaced by the
+    // logger configured in `AddSerilog()` later, once configuration and dependency-injection have both been
+    // set up successfully.
+    Log.Logger = new LoggerConfiguration()
+        .WriteTo.Console()
+        .CreateBootstrapLogger();
+
+    Log.Information("Starting Evaluation Server");
 
     WebApplication.CreateBuilder(args)
-        .ConfigureHost()
         .RegisterServices()
         .Build()
         .SetupMiddleware()
@@ -22,45 +25,6 @@ catch (Exception ex)
 finally
 {
     Log.CloseAndFlush();
-}
-
-return;
-
-void InitializeSerilog()
-{
-    var configuration = new LoggerConfiguration()
-        .MinimumLevel.Override("Microsoft.AspNetCore", LogEventLevel.Warning)
-        .MinimumLevel.Override("Npgsql.Command", LogEventLevel.Warning)
-        .Enrich.FromLogContext()
-        .WriteTo.Console(new CompactJsonFormatter());
-
-    var enableOpenTelemetry = Environment.GetEnvironmentVariable("ENABLE_OPENTELEMETRY");
-    if (enableOpenTelemetry?.ToLower() == "true")
-    {
-        configuration.WriteTo.OpenTelemetry(options =>
-        {
-            options.Endpoint =
-                Environment.GetEnvironmentVariable("OTEL_EXPORTER_OTLP_ENDPOINT") ?? "http://otel-collector:4318";
-            options.Protocol =
-                Environment.GetEnvironmentVariable("OTEL_EXPORTER_OTLP_PROTOCOL") == "grpc"
-                    ? OtlpProtocol.Grpc
-                    : OtlpProtocol.HttpProtobuf;
-
-            options.IncludedData = IncludedData.MessageTemplateTextAttribute
-                                   | IncludedData.TraceIdField
-                                   | IncludedData.SpanIdField;
-            options.BatchingOptions.BatchSizeLimit = 2;
-            options.BatchingOptions.BufferingTimeLimit = TimeSpan.FromSeconds(2);
-            options.BatchingOptions.QueueLimit = 10;
-
-            options.ResourceAttributes = new Dictionary<string, object>
-            {
-                ["service.name"] = "featbit-els"
-            };
-        });
-    }
-
-    Log.Logger = configuration.CreateLogger();
 }
 
 // https://docs.microsoft.com/en-us/aspnet/core/test/integration-tests?view=aspnetcore-6.0#basic-tests-with-the-default-webapplicationfactory
