--- modules/evaluation-server/src/Api/Setup/HostBuilder.cs
@@ -1,13 +1,0 @@
-using Serilog;
-
-namespace Api.Setup;
-
-public static class HostBuilder
-{
-    public static WebApplicationBuilder ConfigureHost(this WebApplicationBuilder builder)
-    {
-        builder.Host.UseSerilog();
-
-        return builder;
-    }
-}
