+++ modules/evaluation-server/src/Api/appsettings.Development.json
@@ -1,4 +1,24 @@
 {
+  "Logging": {
+    "Using": [
+      "Serilog.Sinks.Console"
+    ],
+    "MinimumLevel": {
+      "Default": "Information",
+      "Override": {
+        "Microsoft.AspNetCore": "Warning",
+        "Npgsql.Command": "Information"
+      }
+    },
+    "WriteTo": [
+      {
+        "Name": "Console",
+        "Args": {
+          "outputTemplate": "[{Timestamp:HH:mm:ss} {Level:u3}] {SourceContext}: {Message:lj}{NewLine}{Exception}"
+        }
+      }
+    ]
+  },
   "Kafka": {
     "Producer": {
       "bootstrap.servers": "localhost:29092",
