+++ modules/evaluation-server/src/Api/appsettings.json
@@ -1,4 +1,24 @@
 {
+  "Logging": {
+    "Using": [
+      "Serilog.Sinks.Console"
+    ],
+    "MinimumLevel": {
+      "Default": "Information",
+      "Override": {
+        "Microsoft.AspNetCore": "Warning",
+        "Npgsql.Command": "Warning"
+      }
+    },
+    "WriteTo": [
+      {
+        "Name": "Console",
+        "Args": {
+          "formatter": "Serilog.Formatting.Compact.CompactJsonFormatter, Serilog.Formatting.Compact"
+        }
+      }
+    ]
+  },
   "Kafka": {
     "Producer": {
       "bootstrap.servers": "kafka:9092",
